Fix various msvc compiler warnings. With msvc the compiler flags are explicit passed to nvcc with -Xcompiler.

+++ src/gpu/GridGenerator/StreetPointFinder/StreetPointFinder.cpp
@@ -112,7 +112,7 @@ void StreetPointFinder::prepareSimulationFileData()
 	//////////////////////////////////////////////////////////////////////////
 	// prepare vectors
 
-	uint numberOfCells = this->sparseIndicesLB.size();
+	uint numberOfCells = (uint)this->sparseIndicesLB.size();
 
 	mapNashToConc.resize(numberOfCells);
